organize components into atomic structure
