+++ src/components/molecules/GradientLegend.js
