+++ src/components/molecules/Hint.js
