+++ src/views/index.js
@@ -3,12 +3,10 @@ import { Route, Redirect, Switch } from 'react-router-dom'
 import ExplorerView from './explorer'
 import ColorsView from './colors'
 import withRoot from '../withRoot';
-import ScatterplotView from './scatterplot'
 
 const App = () => (
   <Switch>
     <Route exact path="/" render={() => (<Redirect to="/map/us/counties/avg/all/3.5/38/-97"/>)} />
-    <Route path="/scatterplot" component={ ScatterplotView } />
     <Route exact path="/colors/:highlightedState/:region/:metric/:demographic/:zoom/:lat/:lon/:color?" component={ ColorsView } />
     <Route exact path="/:view/:highlightedState/:region/:metric/:demographic/:zoom/:lat/:lon/:locations?" component={ ExplorerView } />
   </Switch>
